// src/api/generated/model/storyDto.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { StoryDtoItemsItem } from "./storyDtoItemsItem";
import type { StoryDtoType } from "./storyDtoType";
import type { StoryDtoVisibility } from "./storyDtoVisibility";

export interface StoryDto {
  /** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
  id: string;
  title: string;
  /** @nullable */
  subtitle: string | null;
  type: StoryDtoType;
  visibility: StoryDtoVisibility;
  periodStart: string;
  periodEnd: string;
  expiresAt: string;
  /**
   * @nullable
   * @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
   */
  userId: string | null;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  coverImageUrl: string | null;
  items?: StoryDtoItemsItem[];
}